requirements.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.40

# Vector Database
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
tiktoken>=0.5.2

# Utilities
pydantic>=2.5.0

